Split semantic_logger.go into {console,file,mongo}_logger.go
